hh/requirements.txt
asgiref==3.7.2
black==23.7.0
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
